+++ threedi_api_client/files.py
@@ -4,7 +4,7 @@ import logging
 import os
 import re
 from pathlib import Path
-from typing import BinaryIO, Optional, Tuple, Callable
+from typing import BinaryIO, Optional, Tuple, Callable, Union
 from urllib.parse import urlparse
 
 import urllib3
@@ -12,10 +12,10 @@ import urllib3
 from threedi_api_client.openapi import ApiException
 
 CONTENT_RANGE_REGEXP = re.compile(r"^bytes (\d+)-(\d+)/(\d+|\*)$")
-# Ignore timeout settings if the uploaded file is larger than 500MB.
-# This to handle the fact that MinIO computes the MD5sum of the file,
-# which will take longer than a few seconds for larger files.
-UPLOAD_SIZE_NO_TIMEOUT = 500 * 1024 * 1024
+# Default upload timeout has an increased socket read timeout, because MinIO
+# takes very long for completing the upload for larger files. The limit of 10 minutes
+# should accomodate files up to 150 GB.
+DEFAULT_UPLOAD_TIMEOUT = urllib3.Timeout(connect=5.0, read=600.0)
 
 
 logger = logging.getLogger(__name__)
@@ -39,7 +39,7 @@ def download_file(
     url: str,
     target: Path,
     chunk_size: int = 16777216,
-    timeout: float = 5.0,
+    timeout: Optional[Union[float, urllib3.Timeout]] = 5.0,
     pool: Optional[urllib3.PoolManager] = None,
     callback_func: Optional[Callable[[int, int], None]] = None
 ) -> Tuple[Path, int]:
@@ -101,7 +101,7 @@ def download_fileobj(
     url: str,
     fileobj: BinaryIO,
     chunk_size: int = 16777216,
-    timeout: float = 5.0,
+    timeout: Optional[Union[float, urllib3.Timeout]] = 5.0,
     pool: Optional[urllib3.PoolManager] = None,
     callback_func: Optional[Callable[[int, int], None]] = None
 ) -> int:
@@ -186,7 +186,7 @@ def upload_file(
     url: str,
     file_path: Path,
     chunk_size: int = 16777216,
-    timeout: float = 5.0,
+    timeout: Optional[Union[float, urllib3.Timeout]] = None,
     pool: Optional[urllib3.PoolManager] = None,
     md5: Optional[bytes] = None,
     callback_func: Optional[Callable[[int, int], None]] = None
@@ -201,8 +201,8 @@ def upload_file(
         file_path: The file path to read data from.
         chunk_size: The size of the chunk in the streaming upload. Note that this
             function does not do multipart upload. Default: 16MB.
-        timeout: The total timeout in seconds.
-            There is no timeout if the file is larger than 500MB.
+        timeout: The total timeout in seconds. The default is a connect timeout of
+            5 seconds and a read timeout of 10 minutes.
         pool: If not supplied, a default connection pool will be
             created with a retry policy of 3 retries after 1, 2, 4 seconds.
         md5: The MD5 digest (binary) of the file. Supply the MD5 if you already
@@ -259,7 +259,7 @@ def upload_fileobj(
     url: str,
     fileobj: BinaryIO,
     chunk_size: int = 16777216,
-    timeout: float = 5.0,
+    timeout: Optional[Union[float, urllib3.Timeout]] = None,
     pool: Optional[urllib3.PoolManager] = None,
     md5: Optional[bytes] = None,
     callback_func: Optional[Callable[[int, int], None]] = None
@@ -274,8 +274,8 @@ def upload_fileobj(
         fileobj: The (binary) file object to read from.
         chunk_size: The size of the chunk in the streaming upload. Note that this
             function does not do multipart upload. Default: 16MB.
-        timeout: The total timeout in seconds.
-            There is no timeout if the file is larger than 500MB.
+        timeout: The total timeout in seconds. The default is a connect timeout of
+            5 seconds and a read timeout of 10 minutes.
         pool: If not supplied, a default connection pool will be
             created with a retry policy of 3 retries after 1, 2, 4 seconds.
         md5: The MD5 digest (binary) of the file. Supply the MD5 if you already
@@ -322,9 +322,6 @@ def upload_fileobj(
     if file_size == 0:
         raise IOError("The file object is empty.")
 
-    if file_size > UPLOAD_SIZE_NO_TIMEOUT and timeout is not None:
-        timeout = None
-
     if pool is None:
         pool = get_pool()
 
@@ -347,7 +344,7 @@ def upload_fileobj(
         url,
         body=iterable,
         headers=headers,
-        timeout=timeout,
+        timeout=DEFAULT_UPLOAD_TIMEOUT if timeout is None else timeout,
     )
     if response.status != 200:
         raise ApiException(http_resp=response)